Move rigidbody and collider components to proper package
