+++ src/main/java/me/keybarricade/voxeltex/component/collider/ColliderComponentInterface.java
@@ -1,0 +1,3 @@
+package me.keybarricade.voxeltex.component.collider;
+
+public interface ColliderComponentInterface { }
