--- src/main/java/me/keybarricade/voxeltex/physics/collider/ColliderComponentInterface.java
@@ -1,3 +1,0 @@
-package me.keybarricade.voxeltex.physics.collider;
-
-public interface ColliderComponentInterface { }
